Move empty_schema fixture to correct file
